[ins] common ins_init_origin_i_from_flightplan
